pyproject.toml
[tool.poetry]
name = "riff"
version = "0.0.1"
description = "Testing APIs through contracts"
authors = ["jacobsin <jayrad.security@protonmail.com>"]

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.1"
requests = "^2.21"
click = "7.0"
loguru = "^0.2.5"

[tool.poetry.dev-dependencies]
pytest = "^4.4.0"
pytest-cov = "^2.6.1"
pytest-sugar = "^0.9.2"
pytest-xdist = "^1.27.0"

codecov = "^2.0.15"

# formatting
black = { version = "*", allows-prereleases = true }

# docs
jupyter = "^1.0.0"
jupyterlab = "^0.35.0"

# Travis libs
pre-commit = "^1.15.1"

[tool.poetry.scripts]
riff = 'riff.__main__:execute_contract'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

